# docs/api/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API шлюз для доступа к сервисам системы",
        "title": "API Gateway Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/login": {
            "post": {
                "description": "Выполняет вход пользователя в систему",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "security": [
                    {
                        "sessionAuth": []
                    }
                ],
                "description": "Возвращает профиль текущего авторизованного пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Получение профиля пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "sessionAuth": []
                    }
                ],
                "description": "Обновляет данные профиля текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Обновление профиля пользователя",
                "parameters": [
                    {
                        "description": "Новые данные для профиля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProfileUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProfileUpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Создает нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке\nexample: Invalid credentials",
                    "type": "string"
                }
            }
        },
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "description": "Логин пользователя\nrequired: true\nexample: johndoe",
                    "type": "string"
                },
                "password": {
                    "description": "Пароль пользователя\nrequired: true\nexample: securePassword123",
                    "type": "string"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение о результате\nexample: Logged in successfully",
                    "type": "string"
                }
            }
        },
        "models.ProfileResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "description": "Данные пользователя",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
        "models.ProfileUpdateRequest": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "description": "Дата рождения пользователя\nexample: 1990-01-01T00:00:00Z",
                    "type": "string"
                },
                "email": {
                    "description": "Email пользователя\nexample: john.doe@example.com",
                    "type": "string"
                },
                "name": {
                    "description": "Имя пользователя\nexample: John",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "phoneNumber": {
                    "description": "Номер телефона пользователя\nexample: +12345678901",
                    "type": "string"
                },
                "surname": {
                    "description": "Фамилия пользователя\nexample: Doe",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                }
            }
        },
        "models.ProfileUpdateResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Статус операции\nexample: Profile updated successfully",
                    "type": "string"
                },
                "user": {
                    "description": "Обновленные данные пользователя",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
        "models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "login",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email пользователя\nrequired: true\nexample: john.doe@example.com",
                    "type": "string"
                },
                "login": {
                    "description": "Логин пользователя (минимум 3 символа)\nrequired: true\nmin length: 3\nmax length: 50\nexample: johndoe",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "password": {
                    "description": "Пароль пользователя (минимум 6 символов)\nrequired: true\nmin length: 6\nexample: securePassword123",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6
                }
            }
        },
        "models.RegisterResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Статус операции\nexample: user created",
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "description": "Дата рождения пользователя\nexample: 1990-01-01T00:00:00Z",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Дата создания пользователя\nexample: 2023-01-01T12:00:00Z",
                    "type": "string"
                },
                "email": {
                    "description": "Email пользователя\nexample: john.doe@example.com",
                    "type": "string"
                },
                "login": {
                    "description": "Логин пользователя\nexample: johndoe",
                    "type": "string"
                },
                "name": {
                    "description": "Имя пользователя\nexample: John",
                    "type": "string"
                },
                "phoneNumber": {
                    "description": "Номер телефона пользователя\nexample: +12345678901",
                    "type": "string"
                },
                "surname": {
                    "description": "Фамилия пользователя\nexample: Doe",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Дата последнего обновления пользователя\nexample: 2023-01-02T12:00:00Z",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "sessionAuth": {
            "type": "apiKey",
            "name": "user-session",
            "in": "cookie"
        }
    }
}